requirements.txt
coqui-tts           # Text-to-speech
vosk                # Speech recognition

pypinyin            # Chinese pinyin conversion
cutlet              # Japanese to romaji conversion